+++ src/compiler/logic/index.ts
@@ -6,5 +6,5 @@
  */
 
 export * from "./built-ins";
-export * from "./scope-variable-declaration";
+export * from "./scope-declaration";
 export * from "./types";